notiz für mich selber: int = float/2 ist dumm 1/2 != 0 ich bin dumm ich bin dumm ich bin dumm ich bin dumm ich bin dumm ich bin dumm ich bin dumm ich bin dumm ich bin dumm

+++ Magnetfeld Simulator/Wire_parts.cpp
@@ -7,7 +7,7 @@ std::vector<wire_part> Wire_parts::ring()
 {
 	std::vector<wire_part> wire_part_vector = {};
 
-	int middle_x = simulation_size_x / 2, middle_y = simulation_size_y / 2, middle_z = simulation_size_z / 2;
+	float middle_x = simulation_size_x / 2, middle_y = simulation_size_y / 2, middle_z = simulation_size_z / 2;
 
 	for (int current_part = 0; current_part < simulation_wire_parts_count; ++current_part) {
 		wire_part temp = {
@@ -18,6 +18,7 @@ std::vector<wire_part> Wire_parts::ring()
 			(long double)current_part / simulation_wire_parts_count * 2 * PI,
 			(radius * 2 * PI) / simulation_wire_parts_count,
 		};
+		//std::cout << temp.x << "  " << temp.y << "  " << temp.z << "\n";
 		wire_part_vector.push_back(temp);
 	}
